Move nested interface
